refactor: mv xcvr.py transceiver/transceiver.py
preparation to support multiple transceiver type
