+++ src/xcvr_emu/transceiver/__init__.py
@@ -1,0 +1,3 @@
+from .transceiver import CMISTransceiver
+
+__all__ = ['CMISTransceiver']
